feat: scaffold new endgame sections
