+++ feature-flags/feature-flags.development.ts
@@ -20,4 +20,6 @@ export const flagsDevelopment: FeatureFlagsInterface = {
   FEATURE_ECOSYSTEM_ACTORS_TRANSPARENCY_REPORTING: true,
   FEATURE_BUDGET_SUMMARY_PAGE: true,
   FEATURE_ENDGAME_NAVIGATION_SECTION: true,
+  FEATURE_ENDGAME_BUDGET_STRUCTURE_SECTION: true,
+  FEATURE_ENDGAME_BUDGET_TRANSITION_SECTION: true,
 };
