+++ feature-flags/feature-flags.production.ts
@@ -20,4 +20,6 @@ export const flagsProduction: FeatureFlagsInterface = {
   FEATURE_ECOSYSTEM_ACTORS_TRANSPARENCY_REPORTING: true,
   FEATURE_BUDGET_SUMMARY_PAGE: false,
   FEATURE_ENDGAME_NAVIGATION_SECTION: false,
+  FEATURE_ENDGAME_BUDGET_STRUCTURE_SECTION: false,
+  FEATURE_ENDGAME_BUDGET_TRANSITION_SECTION: false,
 };
