+++ src/stories/containers/Endgame/components/BudgetStructureSection/BudgetStructureSection.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import SectionHeader from '../SectionHeader/SectionHeader';
+
+const BudgetStructureSection: React.FC = () => (
+  <div>
+    <SectionHeader
+      title="Endgame Budget Structure"
+      subtitle="Some simple but poignant text about what endgame budgets are about"
+    />
+  </div>
+);
+
+export default BudgetStructureSection;
