+++ src/stories/containers/Endgame/components/KeyChangesSections/KeyChangesSections.tsx
@@ -2,6 +2,7 @@ import styled from '@emotion/styled';
 import lightTheme from '@ses/styles/theme/light';
 import React from 'react';
 import KeyChangeSection from '../KeyChangeSection/KeyChangeSection';
+import KeyChangesBudgetTransitionStatusSection from '../KeyChangesBudgetTransitionStatusSection/KeyChangesBudgetTransitionStatusSection';
 import GovernanceSection from './Sections/GovernanceSection';
 import OperationsSection from './Sections/OperationsSection';
 import TokenUpgradesSection from './Sections/TokenUpgradesSection';
@@ -46,6 +47,8 @@ const KeyChangesSections: React.FC = () => {
           <TokenUpgradesSection />
         </SectionContainer>
       </KeyChangeSection>
+
+      <KeyChangesBudgetTransitionStatusSection />
     </KeyChangesSectionsContainer>
   );
 };
